+++ sync-for-payables/src/codatsyncpayables/models/shared/clientratelimitresetwebhookdata.py
@@ -34,7 +34,7 @@ class ClientRateLimitResetWebhookData:
     > Not all dates from Codat will contain information about time zones.  
     > Where it is not available from the underlying platform, Codat will return these as times local to the business whose data has been synced.
     """
-    quota_remaining: Optional[int] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('QuotaRemaining'), 'exclude': lambda f: f is None }})
+    quota_remaining: Optional[int] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('QuotaRemaining') }})
     r"""Total number of request remaining for your client."""
     reset_reason: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('ResetReason'), 'exclude': lambda f: f is None }})
     r"""The reason for your rate limit quota being reset."""