+++ sync-for-payables/src/codatsyncpayables/sdk.py
@@ -108,7 +108,9 @@ class CodatSyncPayables:
         if client is None:
             client = requests_http.Session()
         
+        
         security_client = utils.configure_security_client(client, security)
+        
         
         if server_url is not None:
             if url_params is not None: